requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.5.0=py38h06a4308_0
argon2-cffi=20.1.0=py38h27cfd23_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py38h0a891b7_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.14.6=py38ha65f79e_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
cryptography=37.0.2=py38h2b5fc30_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h0708190_1
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.3=pyhd8ed1ab_0
imageio=2.21.0=pypi_0
imageio-ffmpeg=0.4.7=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py38h06a4308_0
ipython=8.4.0=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h166bdaf_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py38h06a4308_0
jupyter=1.0.0=py38h06a4308_8
jupyter_client=7.2.2=py38h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py38h06a4308_0
jupyter_server=1.18.1=py38h06a4308_0
jupyterlab=3.4.4=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py38h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hf544144_3
libuuid=1.0.3=h7f8727e_2
libwebp-base=1.2.2=h7f98852_1
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.1.1=py38h7f8727e_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
moviepy=1.0.3=pypi_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py38h06a4308_0
nbconvert=6.4.4=py38h06a4308_0
nbformat=5.3.0=py38h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py38h06a4308_0
nettle=3.6=he412f7d_0
notebook=6.4.12=py38h06a4308_0
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py38ha0e1e83_1
pip=22.1.2=py38h06a4308_0
proglog=0.1.10=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h578d9bd_5
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.12.0=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py38h06a4308_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py38h6a678d5_0
qt=5.9.7=h5867ecd_1
qtconsole=5.3.1=py38h06a4308_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py38h06a4308_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py38h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.5=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py38h06a4308_0
testpath=0.6.0=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.0=py38_cu116
torchvision=0.13.0=py38_cu116
tornado=6.1=py38h27cfd23_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38h06a4308_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38h06a4308_0
xz=5.2.5=h7f8727e_1
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.0=ha95c52a_0
